requirements.txt
# Langchain
langchain 
langchain-google-genai
langchain-community
langchain-cli


# Gemini
google-generativeai


# Environment
python-dotenv


# Flask Frontend/Backend
flask
flask-cors


# Parsers
markdown2 
mistune
html2text
pypdf


# Web Scrapings
serp 
serpapi
googlesearch-python
google-search-results
beautifulsoup4
tavily-python
markdown
requests
lxml


# NLP
langchain-ollama
sentence_transformers


# RAG - Vector Stores
faiss-cpu
pinecone[grpc]>=7.0.0


# Conversational Agents
pyttsx3
speechrecognition
pyaudio

# YouTube Search
youtube-search-python

# Tavily AI Search
tavily-python

